Split the composition-root tests into their own suite
The main Direct Instantiation suite went past the 300-line type-body limit with
the five new cases. The gate needs whole assembler bodies to exercise, so its
tests are the long ones; they get their own file rather than a lint exclusion.
